+++ test/chemistry/test_initial_state_hartree_fock.py
@@ -31,41 +31,41 @@ class TestInitialStateHartreeFock(QiskitChemistryTestCase):
 
     def test_qubits_4_jw_h2(self):
         """ qubits 4 jw h2 test """
-        hrfo = HartreeFock(4, 4, [1, 1], 'jordan_wigner', False)
+        hrfo = HartreeFock(4, [1, 1], 'jordan_wigner', False)
         cct = hrfo.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])
 
     def test_qubits_4_py_h2(self):
         """ qubits 4 py h2 test """
-        hrfo = HartreeFock(4, 4, [1, 1], 'parity', False)
+        hrfo = HartreeFock(4, [1, 1], 'parity', False)
         cct = hrfo.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])
 
     def test_qubits_4_bk_h2(self):
         """ qubits 4 bk h2 test """
-        hrfo = HartreeFock(4, 4, [1, 1], 'bravyi_kitaev', False)
+        hrfo = HartreeFock(4, [1, 1], 'bravyi_kitaev', False)
         cct = hrfo.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])
 
     def test_qubits_2_py_h2(self):
         """ qubits 2 py h2 test """
-        hrfo = HartreeFock(2, 4, 2, 'parity', True)
+        hrfo = HartreeFock(4, 2, 'parity', True)
         cct = hrfo.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.0, 1.0, 0.0, 0.0])
 
     def test_qubits_2_py_h2_cct(self):
         """ qubits 2 py h2 cct test """
-        hrfo = HartreeFock(2, 4, [1, 1], 'parity', True)
+        hrfo = HartreeFock(4, [1, 1], 'parity', True)
         cct = hrfo.construct_circuit('circuit')
         self.assertEqual(cct.qasm(), 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[2];\n'
                                      'u3(pi,0,pi) q[0];\n')
 
     def test_qubits_6_py_lih_cct(self):
         """ qubits 6 py lih cct test """
-        hrfo = HartreeFock(6, 10, [1, 1], 'parity', True, [1, 2])
+        hrfo = HartreeFock(10, [1, 1], 'parity', True, [1, 2])
         cct = hrfo.construct_circuit('circuit')
         self.assertEqual(cct.qasm(), 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[6];\n'
                                      'u3(pi,0,pi) q[0];\n'
@@ -73,7 +73,7 @@ class TestInitialStateHartreeFock(QiskitChemistryTestCase):
 
     def test_qubits_10_bk_lih_bitstr(self):
         """ qubits 10 bk lih bitstr test """
-        hrfo = HartreeFock(10, 10, [1, 1], 'bravyi_kitaev', False)
+        hrfo = HartreeFock(10, [1, 1], 'bravyi_kitaev', False)
         bitstr = hrfo.bitstr
         np.testing.assert_array_equal(bitstr,
                                       [False, False, False, False, True,
@@ -104,7 +104,7 @@ class TestInitialStateHartreeFock(QiskitChemistryTestCase):
 
         qubit_op, _ = core.run(qmolecule)
         qubit_op = op_converter.to_matrix_operator(qubit_op)
-        hrfo = HartreeFock(qubit_op.num_qubits, core.molecule_info['num_orbitals'],
+        hrfo = HartreeFock(core.molecule_info['num_orbitals'],
                            core.molecule_info['num_particles'], mapping.value, False)
         qc = hrfo.construct_circuit('vector')
         hf_energy = qubit_op.evaluate_with_statevector(qc)[0].real + core._nuclear_repulsion_energy